added jwt for verification, route for editing and deleting post and using jwt to verify if the user is the creator

+++ src/app/api/prompt/[id]/route.js
@@ -3,6 +3,11 @@ import connectToDb from "@/utils/database";
 
 //GET POST
 export const GET = async (req, { params }) => {
+  const token = await getToken({ req });
+  if (!token)
+    return new Response("You are not authorized to view this page", {
+      status: 401,
+    });
   try {
     await connectToDb();
     const prompt = await Prompt.findById(params.id);
@@ -16,6 +21,13 @@ export const GET = async (req, { params }) => {
 //EDIT POST
 export const PATCH = async (req, { params }) => {
   const { prompt, tag } = await req.json();
+  const token = await getToken({ req });
+  if (!token)
+    return new Response("You are not authorized to view this page", {
+      status: 401,
+    });
+  if (token.id !== params.id)
+    return new Response("You can only edit your post");
   try {
     await connectToDb();
     const existingPrompt = await Prompt.findById(params.id);
@@ -30,12 +42,19 @@ export const PATCH = async (req, { params }) => {
 
     return new Response(JSON.stringify(existingPrompt), { status: 200 });
   } catch (error) {
-    return new Response("Internal server error", { status: 500 });
+    return new Response("Failed to update the prompt", { status: 500 });
   }
 };
 
 //DELETE POST
 export const DELETE = async (req, { params }) => {
+  const token = await getToken({ req });
+  if (!token)
+    return new Response("You are not authorized to view this page", {
+      status: 401,
+    });
+  if (token.id !== params.id)
+    return new Response("You can only delete your post");
   try {
     await connectToDb();
     await Prompt.findByIdAndRemove(params.id);
